fix/ ветка add-controllers#6 - замечания ревьюера

+++ src/main/java/ru/practicum/shareit/user/service/UserServiceImplementation.java
@@ -1,20 +1,13 @@
 package ru.practicum.shareit.user.service;
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
-import ru.practicum.shareit.exception.InternalErrorException;
+import ru.practicum.shareit.exception.NotFoundException;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.mapper.UserMapper;
-import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.user.storage.UserStorage;
 
-import java.util.Objects;
-
-@Slf4j
 @RequiredArgsConstructor
-@Primary
 @Service
 public class UserServiceImplementation implements UserService {
 
@@ -27,31 +20,19 @@ public class UserServiceImplementation implements UserService {
 
     @Override
     public UserDto createUser(UserDto userDto) {
-        String email = userDto.getEmail();
-        if (userStorage.getEmails().contains(email) || !isValidEmail(email)) {
-            throw new InternalErrorException("Некорректный email");
-        }
-
-        return userStorage.create(userDto);
+        return UserMapper.toUserDto(userStorage.create(userDto));
     }
 
     @Override
     public UserDto updateUser(UserDto userDto, long id) {
-        User updatedUser = userStorage.get(id);
-        String email = userDto.getEmail();
-        if (userStorage.getEmails().contains(email) && !Objects.equals(updatedUser.getEmail(), email)) {
-            throw new InternalErrorException("Некорректный email");
-        }
-
         return userStorage.update(userDto, id);
     }
 
     @Override
     public void deleteUser(long id) {
+        if (userStorage.get(id) == null) {
+            throw new NotFoundException("Пользователь с заданным id не найден");
+        }
         userStorage.delete(id);
-    }
-
-    private static boolean isValidEmail(String email) {
-        return email.matches("^[\\w-\\.]+@[\\w-]+(\\.[\\w-]+)*\\.[a-z]{2,}$");
     }
 }
